+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added knives made from metals
 - Changed recipes for Crossbow and Tripwire Hook
 - Changed recipes for Cartography Table, Fetching Table, Grindstone, Loom, Smithing Table and Smoker
+- Added burn time for valid items like Thatch, Wooden Plate etc.
 
 ## [0.4.5]